Use polars namespace everywhere
