+++ src/cpp/polars/SeriesMask.h
@@ -7,8 +7,10 @@
 
 #include "armadillo"
 
-class Series;
 namespace polars {
+
+    class Series;
+
     // TODO combine Series and SeriesMask via a BaseSeries
     // TODO Create IntSeries and EnumSeries
     // TODO rename Series -> DoubleSeries and SeriesMask -> BoolSeries
